S3 connection working, uploadng to storage. STAC connection working, but publishing of features must be implemented.

+++ s3_connector.py
@@ -1,6 +1,8 @@
 import boto3
 import logging
 
+import botocore.exceptions
+
 import config.s3_config as s3_config
 
 """
@@ -10,14 +12,11 @@ host_base = "https://s3.cl4.du.cesnet.cz"
 use_https = True
 access_key = "your_access_key_should_go_here"
 secret_key = "your_secret_key_should_go_here"
-host_bucket = "copernicus-era5"
+host_bucket = "landsat"
 """
 
 
 class S3Connector:
-    keys = []
-    keys_up_to_date = False
-
     def __init__(
             self,
             logger=logging.getLogger("S3Connector"),
@@ -40,10 +39,7 @@ class S3Connector:
         self.logger.info("Uploading file=" + local_filename + " to S3 as key=" + bucket_key + ".")
 
         # TODO Pokud by bylo paralelizovano, tak zde musi byt nejaky mutex
-        self.keys_up_to_date = False
         self.s3_client.upload_file(local_filename, self.bucket, bucket_key)
-        self.keys.append(bucket_key)
-        self.keys_up_to_date = True
 
     def download_file(self, path_to_download, bucket_key):
         self.logger.info("Downloading key=" + bucket_key + " into file=" + path_to_download + ".")
@@ -54,16 +50,15 @@ class S3Connector:
     def delete_key(self, key):
         self.logger.info("Deleting key=" + key + ".")
 
-        self.keys_up_to_date = False
         self.s3_client.delete_object(Bucket=self.bucket, Key=key)
 
-    def _update_keys(self):
-        """
+    """def _update_keys(self):
+        \"""
         Method updates array of keys which are representing files in S3 storage.
         This is done by using pagination, since the total number of keys one page is able to return is 1000.
 
         :return: Nothing, but method alters (upadtes) array self.keys; and boolean self.keys_up_to_date to True
-        """
+        \"""
         # TODO Pokud by bylo paralelizovano, tak zde musi byt nejaky mutex
         self.keys = []
 
@@ -76,27 +71,31 @@ class S3Connector:
 
         self.keys_up_to_date = True
         # self.logger.info(len(self.keys))
+        """
 
-    def get_contents_keys(self):
-        # TODO Pokud by bylo paralelizovano, tak zde musi byt nejaky mutex
-        if not self.keys_up_to_date:
-            self._update_keys()
+    def check_if_key_exists(self, key, expected_size):
+        try:
+            key_head = self.s3_client.head_object(Bucket=self.bucket, Key=key)
+        except botocore.exceptions.ClientError as e:
+            if e.response['Error']['Code']=="404":
+                # File/key does not exist
+                return False
+            else:
+                # This can be HTTP 403, or other error
+                raise e
 
-        return self.keys
+        # File exists...
 
-    def check_if_key_exists(self, key):
-        if not self.keys_up_to_date:
-            self._update_keys()
-
-        return key in self.get_contents_keys()
-
-    def delete_month(self, year, month):
-        searched_key = year + '/' + month
-        to_be_deleted = [key for key in self.get_contents_keys() if searched_key in key]
-
-        # TODO Pokud by bylo paralelizovano, tak zde musi byt nejaky mutex
-        for key in to_be_deleted:
+        if str(key_head['ContentLength']) == expected_size:
+            # ...and have the right size
+            return True
+        else:
+            # ...but does not have the right size. Let's delete this key and download it again.
             self.delete_key(key)
+            return False
 
-        self.keys_up_to_date = False
-        self._update_keys()
+        
+        """if not self.keys_up_to_date:
+            self._update_keys()
+
+        return key in self.get_contents_keys()"""
